requirements.txt
# Core AI/ML frameworks
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
langchain-community>=0.0.20

# HTTP and API clients
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing and validation
pydantic>=2.5.0
typing-extensions>=4.8.0

# Wikipedia API
wikipedia>=1.4.0

# JSON handling and utilities
jsonschema>=4.20.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Async support
asyncio>=3.4.3

# Rate limiting
tenacity>=8.2.0

# Logging and monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipython>=8.0.0
